# web/openapi.json
{
  "components": {
    "schemas": {
...
      "ApprovalItem": {
        "description": "一条待审 / 已决记录。\n\n``decided_by`` 是 ``user`` 还是 ``auto_approve_A`` / ``auto_approve_AB``，\n决定\"这个月人工放行了几条\"能不能算出来 —— 面板必须显示它，不能只显示状态。",
        "properties": {
          "auto_expire_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Auto Expire At"
          },
          "comment": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Comment"
          },
          "decided_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Decided At"
          },
          "decided_by": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Decided By"
          },
          "grade": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Grade"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "requested_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Requested At"
          },
          "revision_round": {
            "default": 0,
            "title": "Revision Round",
            "type": "integer"
          },
          "score_total": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Score Total"
          },
          "script_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Script Id"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "task_id": {
            "title": "Task Id",
            "type": "string"
          }
        },
        "required": [
          "id",
          "task_id",
          "status"
        ],
        "title": "ApprovalItem",
        "type": "object"
      },
...
      "ApprovalList": {
        "description": "待审 / 已决列表 + 按状态计数（面板顶部的\"待审 N 条\"）。",
        "properties": {
          "approvals": {
            "items": {
              "$ref": "#/components/schemas/ApprovalItem"
            },
            "title": "Approvals",
            "type": "array"
          },
          "counts": {
            "additionalProperties": {
              "type": "integer"
            },
            "title": "Counts",
            "type": "object"
          },
          "limit": {
            "title": "Limit",
            "type": "integer"
          }
        },
        "required": [
          "approvals",
          "counts",
          "limit"
        ],
        "title": "ApprovalList",
        "type": "object"
      },
      "ApprovalResult": {
        "description": "一次决断的结果（**没有** ``ok`` 字段：能返回就是成功）。",
        "properties": {
          "approval_id": {
            "title": "Approval Id",
            "type": "string"
          },
          "comment": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Comment"
          },
          "decision": {
            "title": "Decision",
            "type": "string"
          },
          "grade": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Grade"
          },
          "revision_round": {
            "default": 0,
            "title": "Revision Round",
            "type": "integer"
          },
          "score_total": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Score Total"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "task_id": {
            "title": "Task Id",
            "type": "string"
          }
        },
        "required": [
          "task_id",
          "approval_id",
          "decision",
          "status"
        ],
        "title": "ApprovalResult",
        "type": "object"
      },
      "ApproveBatchBody": {
        "additionalProperties": false,
        "description": "批量通过（原文 §7.2\"批量操作\"）。\n\n路径是 ``/api/v1/approvals/approve_batch`` 而**不是** §04.4.4 表里的\n``/api/v1/tasks/{id}/approve_batch`` —— 后者的 ``{id}`` 在请求体是\n``task_ids`` 的情况下无处可用（T4.4 施工裁定 124，规格书已同步）。",
        "properties": {
          "comment": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Comment"
          },
          "task_ids": {
            "description": "要放行的任务；逐条留痕",
            "items": {
              "type": "string"
            },
            "minItems": 1,
            "title": "Task Ids",
            "type": "array"
          }
        },
        "required": [
          "task_ids"
        ],
        "title": "ApproveBatchBody",
        "type": "object"
      },
      "ApproveBody": {
        "additionalProperties": false,
        "description": "确认放行（``awaiting_approval → queued_voice``）。",
        "properties": {
          "comment": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Comment"
          }
        },
        "title": "ApproveBody",
        "type": "object"
      },
...
      "BatchFailure": {
        "description": "批量里失败的那一条（**不**因为一条失败就回滚其余的）。",
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "remediation": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Remediation"
          },
          "task_id": {
            "title": "Task Id",
            "type": "string"
          }
        },
        "required": [
          "task_id",
          "code",
          "message"
        ],
        "title": "BatchFailure",
        "type": "object"
      },
      "BatchResult": {
        "description": "批量通过的逐条结果。\n\n``approved`` / ``failed`` 都如实返回：批量是\"尽力而为\"的人工操作，\n把 9 条成功 + 1 条失败报成\"整体失败\"，会让人重按一次 —— 而重复放行\n撞上 ``APPROVAL_NOT_PENDING`` 只是噪音，真正的问题是没人知道哪 9 条成了。",
        "properties": {
          "approved": {
            "items": {
              "$ref": "#/components/schemas/ApprovalResult"
            },
            "title": "Approved",
            "type": "array"
          },
          "failed": {
            "items": {
              "$ref": "#/components/schemas/BatchFailure"
            },
            "title": "Failed",
            "type": "array"
          },
          "requested": {
            "title": "Requested",
            "type": "integer"
          }
        },
        "required": [
          "requested",
          "approved",
          "failed"
        ],
        "title": "BatchResult",
        "type": "object"
      },
...
      "DiscardBody": {
        "additionalProperties": false,
        "description": "放弃（``awaiting_approval → discarded``）。可被 :class:`RescueBody` 捞回。",
        "properties": {
          "reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          }
        },
        "title": "DiscardBody",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "RejectBody": {
        "additionalProperties": false,
        "description": "退回改稿（``awaiting_approval → editing``）。``comment`` **必填**。\n\n这里的 ``min_length=1`` 只挡\"完全没填\"这个最常见的错；\"只填了空格\"由服务层\n的 ``APPROVAL_COMMENT_REQUIRED`` 拒绝 —— 业务不变量只有一个权威落点，\nREST 层不抄第二遍语义（§04.4.4 不变量 3）。",
        "properties": {
          "comment": {
            "description": "必填：下一轮改稿只认这条意见",
            "minLength": 1,
            "title": "Comment",
            "type": "string"
          }
        },
        "required": [
          "comment"
        ],
        "title": "RejectBody",
        "type": "object"
      },
...
      "RescueBody": {
        "additionalProperties": false,
        "description": "捞回（``discarded`` / ``canceled`` → ``pending``）。",
        "properties": {
          "reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          }
        },
        "title": "RescueBody",
        "type": "object"
      },
      "RescueResult": {
        "description": "一次捞回的结果。",
        "properties": {
          "previous_status": {
            "title": "Previous Status",
            "type": "string"
          },
          "revision_round": {
            "default": 0,
            "title": "Revision Round",
            "type": "integer"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "task_id": {
            "title": "Task Id",
            "type": "string"
          },
          "topic_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Topic Id"
          },
          "topic_restored": {
            "default": false,
            "title": "Topic Restored",
            "type": "boolean"
          }
        },
        "required": [
          "task_id",
          "status",
          "previous_status"
        ],
        "title": "RescueResult",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "AI 全自动营销号制片台",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/approvals": {
      "get": {
        "description": "待审 / 已决列表（``counts`` 是面板顶部的\"待审 N 条\"）。",
        "operationId": "list_approvals_api_v1_approvals_get",
        "parameters": [
          {
            "description": "默认只看待审",
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "default": "pending",
              "description": "默认只看待审",
              "pattern": "^(pending|approved|rejected|discarded|expired)$",
              "title": "Status",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 200,
              "maximum": 500,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApprovalList"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Approvals",
        "tags": [
          "approvals"
        ]
      }
    },
    "/api/v1/approvals/approve_batch": {
      "post": {
        "description": "批量通过：**逐条**决断 + 逐条写 ``audit_ops``，一条失败不影响其余。",
        "operationId": "approve_batch_api_v1_approvals_approve_batch_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApproveBatchBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BatchResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Approve Batch",
        "tags": [
          "approvals"
        ]
      }
    },
...
    "/api/v1/tasks/{task_id}/approve": {
      "post": {
        "description": "确认放行：``awaiting_approval → queued_voice``。",
        "operationId": "approve_task_api_v1_tasks__task_id__approve_post",
        "parameters": [
          {
            "in": "path",
            "name": "task_id",
            "required": true,
            "schema": {
              "title": "Task Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/ApproveBody"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Body"
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApprovalResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Approve Task",
        "tags": [
          "approvals"
        ]
      }
    },
    "/api/v1/tasks/{task_id}/discard": {
      "post": {
        "description": "放弃：``awaiting_approval → discarded``（可经 ``/rescue`` 捞回）。",
        "operationId": "discard_task_api_v1_tasks__task_id__discard_post",
        "parameters": [
          {
            "in": "path",
            "name": "task_id",
            "required": true,
            "schema": {
              "title": "Task Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/DiscardBody"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Body"
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApprovalResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Discard Task",
        "tags": [
          "approvals"
        ]
      }
    },
    "/api/v1/tasks/{task_id}/reject": {
      "post": {
        "description": "退回改稿：``awaiting_approval → editing``（``revision_round + 1``，意见必填）。",
        "operationId": "reject_task_api_v1_tasks__task_id__reject_post",
        "parameters": [
          {
            "in": "path",
            "name": "task_id",
            "required": true,
            "schema": {
              "title": "Task Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RejectBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApprovalResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Reject Task",
        "tags": [
          "approvals"
        ]
      }
    },
    "/api/v1/tasks/{task_id}/rescue": {
      "post": {
        "description": "捞回：``discarded`` / ``canceled`` → ``pending``（误点放弃的唯一补救）。",
        "operationId": "rescue_task_api_v1_tasks__task_id__rescue_post",
        "parameters": [
          {
            "in": "path",
            "name": "task_id",
            "required": true,
            "schema": {
              "title": "Task Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/RescueBody"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Body"
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RescueResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Rescue Task",
        "tags": [
          "approvals"
        ]
      }
    },
...
  }
}